#849 Moved utility classes to subpackage utils.
